Script to monitor Services post-reboot
Made using ChatGPT4. Displays and monitors specific named services for run status and waits until all are running before closing. Script is designed to be modified once and then just run as needed as opposed to taking arguments.

Some services utilize a $ in their name, which you will need to change the double quotes to single quotes in the array.

Originally made to check for services or processes, but I dropped the processes part because I ended up not needing to check for those. May redesign to check for either or both.